OneDrive/Desktop/Ai-proj/knowledge_base_agent/requirements.txt
# For CUDA support, please install the appropriate version of torch, torchvision, and torchaudio
# See https://pytorch.org/get-started/locally/
streamlit
faiss-cpu # or faiss-gpu for CUDA
pypdf
sentence-transformers
transformers
torch
torchvision
torchaudio
